bipeds.h: biped_approximate_surface_index returns long (ai_script: 1 function strict exact)
Header packet P4 (150K lane). January's only external caller, ai_script.obj, tests the result with a full 32-bit cmp eax,-1 (no sign extension); the short prototype produced cmp ax,0xffff. The owner prototype and the bipeds.c definition now return long and the unevidenced (short) cast is dropped; declaration count unchanged, callee bytes unchanged. _ai_scripting_teleport_starting_location_private (288) becomes strict exact with an untouched body, so its park (premise: the widening perturbs consumers) is removed.

Consumers: 23 bipeds.h includers, shadow-compiled H vs P (0 losses). Live tree: full ninja + stable_verdicts sweep +1 gained / 0 regressions; parks 290/0/0.

+++ source/units/bipeds.c
@@ -1492,11 +1492,11 @@ static long biped_find_ground_surface(
 	return surface_index;
 }
 
-short biped_approximate_surface_index(
+long biped_approximate_surface_index(
 	long biped_index,
 	real_point3d *point)
 {
-	return (short)biped_find_ground_surface(biped_index, global_down3d, 2.f, point, NULL);
+	return biped_find_ground_surface(biped_index, global_down3d, 2.f, point, NULL);
 }
 
 long biped_find_pathfinding_surface_index(
